Enforce sections in student exam-taking: timers, forward-only navigation, negative marking
Students previously saw section metadata nowhere in the actual attempt flow.
This adds real per-section countdown timers with server-tracked deadlines,
Free/Sequential/Locked navigation enforcement (client + server gating on
both question- and section-level movement), and negative marking scoring
(section-level overriding exam-level, decimal-precise, floored at zero).

Fixes a bug caught during manual testing where the forward-only navigation
guard in TakeExam's goToIndex blocked the normal one-step "Save & Next"
advance, stranding students on the first question of any Sequential/Locked
section.

+++ Backend/Services/ResultService/OnlineExamSystem.Result.Domain/AdminAttemptResult.cs
@@ -13,7 +13,7 @@ public class AdminAttemptResult
 {
     public Guid AttemptId { get; init; }
     public Guid UserId { get; init; }
-    public int TotalScore { get; init; }
+    public decimal TotalScore { get; init; }
     public bool Passed { get; init; }
     public DateTime SubmittedAtUtc { get; init; }
     public IReadOnlyList<QuestionResult> Questions { get; init; } = [];
