+++ Backend/Services/ResultService/OnlineExamSystem.Result.Domain/ExamResultSummary.cs
@@ -5,7 +5,7 @@ public class ExamResultSummary
     public Guid AttemptId { get; init; }
     public Guid ExamId { get; init; }
     public string ExamTitle { get; init; } = string.Empty;
-    public int TotalScore { get; init; }
+    public decimal TotalScore { get; init; }
     public int TotalMarks { get; init; }
     public int PassingMarks { get; init; }
     public bool Passed { get; init; }
@@ -22,7 +22,7 @@ public class QuestionResult
     public Guid QuestionId { get; init; }
     public string QuestionText { get; init; } = string.Empty;
     public int Marks { get; init; }
-    public int MarksAwarded { get; init; }
+    public decimal MarksAwarded { get; init; }
     public Guid? SelectedOptionId { get; init; }
     public bool IsCorrect { get; init; }
     public IReadOnlyList<QuestionResultOption> Options { get; init; } = [];
